pipelines/rj_iplanrio__data_catalog/pyproject.toml
[project]
name = "rj_iplanrio__data_catalog"
version = "0.1.0"
description = "Pipeline data_catalog da secretaria iplanrio"
dependencies = [
    "google-cloud-bigquery>=3.29.0",
    "google-cloud-storage>=2.10.0",
    "polars>=1.35.2",
]
